add resolver tracing metrics into extensions object for apollo-server queries
